# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitOS API",
    "description": "Deterministic fitness engine API. Every number is computed in packages/core, never by a language model (spec §19.1).",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Required on every /v1 route except those marked public (§10)."
      }
    },
...
  },
  "paths": {
...
    "/user/goal": {
      "get": {
        "summary": "Active goal with current targets",
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "goal": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "goalType": {
                          "type": "string",
                          "enum": [
                            "muscle-gain",
                            "fat-loss",
                            "recomposition",
                            "strength",
                            "general",
                            "maintenance"
                          ]
                        },
                        "targetWeightKg": {
                          "type": "number",
                          "nullable": true
                        },
                        "startedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "goalType",
                        "targetWeightKg",
                        "startedAt"
                      ],
                      "additionalProperties": false
                    },
                    "targets": {
                      "type": "object",
                      "properties": {
                        "effectiveFrom": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        "kcal": {
                          "type": "integer"
                        },
                        "proteinG": {
                          "type": "integer"
                        },
                        "carbG": {
                          "type": "integer"
                        },
                        "fatG": {
                          "type": "integer"
                        },
                        "fiberG": {
                          "type": "integer"
                        },
                        "bmr": {
                          "type": "integer"
                        },
                        "tdeeEstimate": {
                          "type": "integer"
                        },
                        "rationale": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "reason": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "effectiveFrom",
                        "kcal",
                        "proteinG",
                        "carbG",
                        "fatG",
                        "fiberG",
                        "bmr",
                        "tdeeEstimate",
                        "rationale",
                        "reason"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "goal",
                    "targets"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "VALIDATION_FAILED",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "UPSTREAM_UNAVAILABLE",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "issue"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "VALIDATION_FAILED",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "UPSTREAM_UNAVAILABLE",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "issue"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "VALIDATION_FAILED",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "UPSTREAM_UNAVAILABLE",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "issue"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace the active goal; closes the old one and recomputes targets",
        "tags": [
          "user"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "goalType": {
                    "type": "string",
                    "enum": [
                      "muscle-gain",
                      "fat-loss",
                      "recomposition",
                      "strength",
                      "general",
                      "maintenance"
                    ]
                  },
                  "targetWeightKg": {
                    "type": "number",
                    "minimum": 30,
                    "maximum": 300,
                    "nullable": true
                  }
                },
                "required": [
                  "goalType"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "goal": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "goalType": {
                          "type": "string",
                          "enum": [
                            "muscle-gain",
                            "fat-loss",
                            "recomposition",
                            "strength",
                            "general",
                            "maintenance"
                          ]
                        },
                        "targetWeightKg": {
                          "type": "number",
                          "nullable": true
                        },
                        "startedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "goalType",
                        "targetWeightKg",
                        "startedAt"
                      ],
                      "additionalProperties": false
                    },
                    "targets": {
                      "type": "object",
                      "properties": {
                        "effectiveFrom": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        "kcal": {
                          "type": "integer"
                        },
                        "proteinG": {
                          "type": "integer"
                        },
                        "carbG": {
                          "type": "integer"
                        },
                        "fatG": {
                          "type": "integer"
                        },
                        "fiberG": {
                          "type": "integer"
                        },
                        "bmr": {
                          "type": "integer"
                        },
                        "tdeeEstimate": {
                          "type": "integer"
                        },
                        "rationale": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "reason": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "effectiveFrom",
                        "kcal",
                        "proteinG",
                        "carbG",
                        "fatG",
                        "fiberG",
                        "bmr",
                        "tdeeEstimate",
                        "rationale",
                        "reason"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "goal",
                    "targets"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "VALIDATION_FAILED",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "UPSTREAM_UNAVAILABLE",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "issue"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "VALIDATION_FAILED",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "UPSTREAM_UNAVAILABLE",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "issue"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "VALIDATION_FAILED",
                            "CONFLICT",
                            "RATE_LIMITED",
                            "UPSTREAM_UNAVAILABLE",
                            "INTERNAL"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "path",
                              "issue"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Versioned base path"
    }
  ]
}